requirements.txt
api==0.0.7
cryptography==3.4.8
Flask==2.0.1
Flask_Login==0.6.2
Flask_Session==0.4.0
Flask_WTF==1.1.1
pymongo==4.3.3
python-dotenv==1.0.0
redis==4.5.1
termcolor==2.2.0
WTForms==3.0.1
gunicorn==20.1.0


